requirements.txt
streamlit
pandas
numpy
requests
azure.cognitiveservices.speech
openpyxl
wave
openpyxl
numpy